Fix incorrect exit code successful run

+++ bls-runtime/src/main.rs
@@ -167,7 +167,7 @@ async fn wasm_runtime(mut cfg: CliConfig, cli_command_opts: CliCommandOpts) -> C
 
     let exit_status = blockless_run(cfg.0).await;
     info!("The wasm execute finish, the exit code: {}", exit_status.code);
-    (exit_status.code as u8).into()
+    exit_status.code.into()
 }
